Remove unnecessary files
The strategy has generally changed for hosting the catui
load balancer, and it's generally not very useful to have
these files anymore. Now a desktop app or service is going
to host a catui server. The `catuid` program doesn't really
fit this model. The reason we're going this direction is so
we can play more nicely with the platforms that already
exist and build on top of them instead of against them. Users
will benefit by feeling like apps are native to the desktop
environment and won't even need to know they're a separate
implementation under the hood.